Added tests fro archive export
