+++ core/api/tests/test_cp_archive.py
@@ -2,8 +2,6 @@ import pytest
 from django.urls import reverse
 
 from core.api.tests.base import BaseTest
-from core.api.tests.factories import CPRaportFormatRowFactory
-from core.models import AdmRecordArchive
 from core.models.country_programme import CPReport
 from core.models.country_programme_archive import CPReportArchive
 
@@ -64,89 +62,6 @@ class TestVersionsList(BaseTest):
 
         response = self.client.get(self.url, {"country_id": cp_report_2019.country_id})
         assert response.status_code == 400
-
-
-@pytest.fixture(name="_setup_old_version_2019")
-def setup_old_version_2019(cp_report_2019, substance, blend, time_frames, user):
-    cp_report_2019.status = CPReport.CPReportStatus.FINAL
-    cp_report_2019.version = 2
-    cp_report_2019.save()
-
-    CPRaportFormatRowFactory.create(
-        blend=blend,
-        substance=None,
-        section="B",
-        time_frame=time_frames[(2000, None)],
-        sort_order=2,
-    )
-
-    CPRaportFormatRowFactory.create(
-        blend=None,
-        substance=substance,
-        section="A",
-        time_frame=time_frames[(2000, None)],
-        sort_order=1,
-    )
-
-    cp_ar = CPReportArchive.objects.create(
-        name=cp_report_2019.name,
-        year=cp_report_2019.year,
-        country=cp_report_2019.country,
-        status=cp_report_2019.status,
-        version=1,
-        created_by=user,
-        last_updated_by=user,
-    )
-
-    return cp_ar
-
-
-@pytest.fixture(name="_setup_old_version_2005")
-def setup_old_version_2005(
-    cp_report_2005, substance, blend, adm_rows, adm_columns, time_frames, user
-):
-    cp_report_2005.status = CPReport.CPReportStatus.FINAL
-    cp_report_2005.version = 2
-    cp_report_2005.save()
-
-    CPRaportFormatRowFactory.create(
-        blend=blend,
-        substance=None,
-        section="B",
-        time_frame=time_frames[(2000, None)],
-        sort_order=2,
-    )
-
-    CPRaportFormatRowFactory.create(
-        blend=None,
-        substance=substance,
-        section="A",
-        time_frame=time_frames[(2000, None)],
-        sort_order=1,
-    )
-
-    cp_ar = CPReportArchive.objects.create(
-        name=cp_report_2005.name,
-        year=cp_report_2005.year,
-        country=cp_report_2005.country,
-        status=cp_report_2005.status,
-        version=1,
-        created_by=user,
-        last_updated_by=user,
-    )
-
-    adm_b_row = adm_rows[0]
-    adm_b_col = adm_columns[0]
-
-    AdmRecordArchive.objects.create(
-        row=adm_b_row,
-        column=adm_b_col,
-        section="B",
-        value_text="Treviso",
-        country_programme_report=cp_ar,
-    )
-
-    return cp_ar
 
 
 class TestGetOldVersion(BaseTest):
